pyproject.toml
[tool.poetry]
name = "data-pipeline-demo"
version = "0.1.0"
description = "Data Pipeline Demo - Technical Proof of Concept"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "src" },
    { include = "local" },
]

[tool.poetry.scripts]
pipeline-local = "local.cli:main"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
apache-airflow = "2.10.5"
snowflake-connector-python = "^3.14.0"
snowflake-sqlalchemy = "^1.7.3"
pandas = "^2.1.1"
numpy = "1.26.4"
pyarrow = "14.0.1"
boto3 = "^1.34.0"
s3fs = "^2025.3.0"
python-dotenv = "^1.1.0"
psycopg2-binary = "2.9.10"
sqlalchemy = "1.4.47"
tqdm = "4.67.1"
psutil = "^5.9.0"
openpyxl = "3.1.4"
duckdb = "0.10.1"
geopy = "^2.4.0"
db-dtypes = "1.4.2"
typer = "^0.15.0"
click = "^8.1.0"
rich = "^13.7.0"
cryptography = "^41.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
isort = "^5.13.0"
ruff = "^0.8.0"
pytest = "^8.0.0"
pytest-cov = "^6.0.0"
ipython = "^8.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 150
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 150
known_first_party = ["src", "local", "dags"]

[tool.ruff]
line-length = 150
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict parameter
    "B018",   # useless expression (valid pattern in Airflow DAGs)
    "W291",   # trailing whitespace (common in SQL strings)
    "B904",   # raise without from (too noisy for existing codebase)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["B011", "F841", "B007"]
"local/cli.py" = ["E402"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning:sqlalchemy.*:",
    "ignore::DeprecationWarning:airflow.*:",
]
log_cli = false
log_cli_level = "CRITICAL"
